statebucket/function-source/requirements.txt
Flask==2.1.1
werkzeug==2.0.3
google-cloud-firestore==2.4.0